Improve locking around configmap loading.  Delete binary.
Improve locking around the configmap loading.  There was a lock acquired
when the configmap is being used to determine identity, if the map was
not initialized there is a call to the API server while the lock is
being held.  This causes large amounts of latency when it is not
present.

We now attempt to load/reload the configmap from a single go routine and
only acquire read locks when using the configmap for identity calls.

Delete the compiled binary that was accidently checked in.